moved test to integration test folder
